# api/userProgram/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a swag test http server.",
        "title": "USER WEB API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7001",
    "basePath": "/api/v1",
    "paths": {
...
        "/user/center/get": {
            "get": {
                "description": "获取用户个人中心信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取用户个人中心信息",
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controller.UserCenterInfoModel"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "controller.UserCenterInfoModel": {
            "type": "object",
            "properties": {
                "orderWaitCount": {
                    "description": "待完成订单",
                    "type": "integer"
                },
                "orderWaitPayCount": {
                    "description": "待支付订单",
                    "type": "integer"
                },
                "productCollectionCount": {
                    "description": "收藏商品",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}